requirements.txt
requests
bs4
collection
psycopg2-binary
datetime
pytest
